fix: packages build setup
