+++ packages/core/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@nosto/search-js-preact",
+  "name": "@nosto/search-js-core",
   "version": "0.1.0",
   "license": "ISC",
   "type": "module",
@@ -7,11 +7,5 @@
   "scripts": {
     "build": "vite build",
     "test": "vitest --run"
-  },
-  "peerDependencies": {
-    "preact": "^10.26.2"
-  },
-  "devDependencies": {
-    "@testing-library/preact": "^3.2.4"
   }
 }
