+++ packages/currencies/index.ts
