+++ packages/thumbnails/src/types.ts
@@ -1,5 +1,6 @@
 /**
  * Nosto thumbnail size code.
+ * Sometimes referred to as `imageVersion`.
  *
  * * 1: 170x170 px
  * * 2: 100x100 px
@@ -15,7 +16,7 @@
  * * 12: 400x400 px (Square)
  * * 13: 750x750 px (Square)
  */
-export type ThumbnailSize = "1" | "2" | "3" | "4" | "5" | "6" | "7" | "8" | "9" | "10" | "11" | "12" | "13" | "orig"
+export type NostoSize = "1" | "2" | "3" | "4" | "5" | "6" | "7" | "8" | "9" | "10" | "11" | "12" | "13" | "orig"
 
 /**
  * Shopify thumbnail size code.
@@ -42,3 +43,19 @@ export type ShopifySize =
   | "grande"
   | "original"
   | "master"
+
+/**
+ * Shopify compatible Nosto thumbnail size code.
+ * This is a subset of {@link NostoSize} that is compatible with Shopify CDN.
+ *
+ * * 1: 170x170 px
+ * * 2: 100x100 px
+ * * 3: 90x70 px
+ * * 4: 50x50 px
+ * * 5: 30x30 px
+ * * 6: 100x140 px
+ * * 7: 200x200 px
+ * * 8: 400x400 px
+ * * 9: 750x750 px
+ */
+export type NostoShopifySize = "1" | "2" | "3" | "4" | "5" | "6" | "7" | "8" | "9"
