+++ tsconfig.json
@@ -25,5 +25,5 @@
     "jsxFactory": "h",
     "jsxFragmentFactory": "Fragment"
   },
-  "exclude": ["**/vite.config.ts", "**/test"]
+  "exclude": ["dist", "**/vite.config.ts", "**/test"]
 }
